adding lombok, changing date to localdatetime and fixing some relations

+++ src/main/java/com/syst/trades/repository/StatusRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.syst.trades.model.Status;
 
-public interface StatusRepository extends JpaRepository<Status, Long>{
+public interface StatusRepository extends JpaRepository<Status, Integer>{
 
 }
